% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{bs.nhpp.time}
\alias{bs.nhpp.time}
\title{bootstrap of NHPP}
\usage{
bs.nhpp.time(intensity, range, maximum)
}
\arguments{
\item{srm}{An object of an estimated result by Rsrat}
}
\value{
A time data (time series)
}
\description{
Generate time data (bootstrap samples) from a given NHPP.
}
