% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{bs.srm.group}
\alias{bs.srm.group}
\title{bootstrap of NHPP}
\usage{
bs.srm.group(srm, data)
}
\arguments{
\item{srm}{An object of model provided from Rsrat}

\item{data}{An object of faultdata}
}
\value{
An object of faultdata of bootstrap sample
}
\description{
Generate group data (bootstrap samples) from a given NHPP.
}
