% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{bs.srm.time}
\alias{bs.srm.time}
\title{bootstrap of NHPP}
\usage{
bs.srm.time(srm, data, maximum)
}
\arguments{
\item{srm}{An object of model provided from Rsrat}

\item{data}{An object of faultdata}

\item{maximum}{A value of the maximum value of intensity for an interval (0, the last observation time)}
}
\value{
An object of faultdata of bootstrap sample
}
\description{
Generate time data (bootstrap samples) from a given NHPP.
}
