% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.faultdata}
\alias{check.faultdata}
\title{Check the category of fault data}
\usage{
check.faultdata(data)
}
\arguments{
\item{data}{An object of faultdata}
}
\value{
A string of "time", "group" and "unknown"
}
\description{
This is a function to decide whether a given fault data is time data or group data.
}
