% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks_mpfr.R
\name{comp.Pdminus.mp}
\alias{comp.Pdminus.mp}
\title{Compute P(D->d-) for the generalized KS statistic}
\usage{
comp.Pdminus.mp(ctime, count, dminus, cdf, prec = 128)
}
\arguments{
\item{ctime}{A sequence represents time slots (bins)}

\item{count}{A sequence indicates the number of samples falls int a bin}

\item{dminus}{A value of d-}

\item{cdf}{A function of CDF. It is allowd to get a closure.}

\item{prec}{An integer to indicate the precision}
}
\value{
A value of the probability
}
\description{
Compute P(D->d-) for the generalized KS statistic
}
