% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_Pdminus_arma}
\alias{compute_Pdminus_arma}
\title{Compute P(D->d-) for the generalized KS statistic}
\usage{
compute_Pdminus_arma(ctime, count, dminus, cdf, solve)
}
\arguments{
\item{ctime}{A sequence represents time slots (bins)}

\item{count}{A sequence indicates the number of samples falls int a bin}

\item{dminus}{A value of d-}

\item{cdf}{A function of CDF. It is allowd to get a closure.}

\item{solve}{A function to solve a linear equation.}
}
\value{
A value of the probability
}
\description{
This is an experimental function instead of compute_Pdminus.
This is not used yet.
}
