% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_Pdplus_arma}
\alias{compute_Pdplus_arma}
\title{Compute P(D+>d+) for the generalized KS statistic}
\usage{
compute_Pdplus_arma(ctime, count, dplus, cdf, solve)
}
\arguments{
\item{ctime}{A sequence represents time slots (bins)}

\item{count}{A sequence indicates the number of samples falls int a bin}

\item{dplus}{A value of d+}

\item{cdf}{A function of CDF. It is allowd to get a closure.}
}
\value{
A value of the probability
}
\description{
This is an experimental function instead of compute_Pdplus.
This is not used yet.
}
