% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{eic}
\alias{eic}
\title{EIC for SRM}
\usage{
eic(result, bsample = 100, alpha = 0.95)
}
\arguments{
\item{srm}{An object of an estimated result by Rsrat}
}
\value{
resA time data (time series)
}
\description{
Compute EIC for SRM
}
