% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{eic.srm.sample}
\alias{eic.srm.sample}
\title{EIC for SRM}
\usage{
eic.srm.sample(obj, bsample = 100)
}
\arguments{
\item{obj}{An object of an estimated result by Rsrat}

\item{bsample}{An integer of the number of bootstrap samples}
}
\value{
A vector of samples
}
\description{
Generate samples of EIC
}
