% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstest.R
\name{gks.srm.test.mc2}
\alias{gks.srm.test.mc2}
\title{Generalized KS test for software reliability models with Monte-Carlo testing
under unknown model parameters}
\usage{
gks.srm.test.mc2(
  obj,
  alternative = c("two.sided", "less", "greater"),
  b = NA,
  alpha = 0.01,
  cv = 0.1,
  lhc = FALSE,
  seed = NA
)
}
\arguments{
\item{obj}{An object of an estimated result by Rsrat}

\item{alternative}{A string indicates the alternative hypothesis and
must be one of "two.sided" (default), "less", or "greater".}

\item{b}{An integer for the number of resamples. If it is NULL, the number of resamples
is determined from the significant level and the coefficient variation.}

\item{alpha}{A value of the significant level which is used to determine the number
of resamples. The default is 0.01. If \code{b} is not NULL, this is ignored.}

\item{cv}{A value of the coefficient variation which is used to determine the number
of resamples. The default is 0.1. If \code{b} is not NULL, this is ignored.}

\item{lhc}{A logical indicates resamples are drawn by Latin hypercude.}

\item{seed}{An integer for the seed of random numbers. If it is NULL, the seed is not set.}
}
\value{
A list with components;
\item{statistic}{A value of the test statistic.}
\item{p.value}{A value of the p-value of the test.}
\item{alternative}{A string of the alternative hypothesis.}
\item{b}{An integer for the number of resamples.}
\item{alpha}{A value of the significant level.}
\item{cv}{A value of the coefficient variation.}
\item{lhc}{A logical indicates resamples are drawn by Latin hypercude.}
\item{seed}{An integer for the seed of random numbers.}
}
\description{
Perform the generalized KS test for the estimated software reliability models
with Monte-Carlo testing
}
\examples{
data(dacs)
result <- fit.srm.nhpp(fault=sys1g[1:30], srm.name=c("exp"))
gks.srm.test.mc2(result)
}
