% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks_mpfr.R
\name{gks.srm.test.mp}
\alias{gks.srm.test.mp}
\title{Generalized KS test for software reliability models}
\usage{
gks.srm.test.mp(srm, alternative = c("two.sided", "less", "greater"),
  prec = 128)
}
\arguments{
\item{srm}{An object for the estimated software reliability model by Rsrat}

\item{alternative}{A string indicates the alternative hypothesis and
must be one of "two.sided" (default), "less", or "greater".}

\item{prec}{An integer to indicate the precision}
}
\value{
A list with components;
\item{statistic}{A value of the test statistic.}
\item{p.value.lower}{A lower bound of the p-value of the test.}
\item{p.value.upper}{A upper bound of the p-value of the test.}
\item{alternative}{A string of the alternative hypothesis.}
}
\description{
Peform the generalized KS test for the estimated software reliability models
}
\examples{
data(dacs)
result <- fit.srm.nhpp(fault=sys1g[1:30], srm.name=c("exp"))
gks.test.mp(result$srm)
}
