% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstest.R
\name{gks.srm.test}
\alias{gks.srm.test}
\title{Generalized KS test for software reliability models with group data}
\usage{
gks.srm.test(obj, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{obj}{An object of an estimated result by Rsrat}

\item{alternative}{A string indicates the alternative hypothesis and
must be one of "two.sided" (default), "less", or "greater".}
}
\value{
A list with components;
\item{statistic}{A value of the test statistic.}
\item{p.value.lower}{A lower bound of the p-value of the test.}
\item{p.value.upper}{A upper bound of the p-value of the test.}
\item{alternative}{A string of the alternative hypothesis.}
}
\description{
Perform the generalized KS test for the estimated software reliability models
}
