% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gof4srm}
\alias{gof4srm}
\title{gof4srm: Goodness-of-fit tests for software reliability models}
\description{
This is a package of goodness-of-fit tests for software reliability (growth)
models. This provides Kolmogorov-Smirnov test for both time and grouped data,
and EIC (extended information criterion).This package provides estimation
programs for software reliability growth
}
