% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstest.R
\name{ks.srm.test}
\alias{ks.srm.test}
\title{KS test for software reliability models}
\usage{
ks.srm.test(
  obj,
  alternative = c("two.sided", "less", "greater"),
  method = c("mc", "gks", "mc2", "yamada"),
  ...
)
}
\arguments{
\item{obj}{An object of an estimated result by Rsrat}

\item{alternative}{A string indicates the alternative hypothesis and
must be one of "two.sided" (default), "less", or "greater".}

\item{method}{A string indicates the method for KS test with group data. "mc"
and "mc2" are Monte-Carlo methods. "gks" is the generalized KS test. "yamada"
is an approximation. The default is "mc".}

\item{...}{A list of other options that are passed to concrete functions.}
}
\value{
A list with components;
\item{statistic}{A value of the test statistic.}
\item{p.value}{A value of the p-value of the test.}
\item{p.value.lower}{A lower bound of the p-value of the test. (gks)}
\item{p.value.upper}{A upper bound of the p-value of the test. (gks)}
\item{alternative}{A string of the alternative hypothesis.}
\item{b}{An integer for the number of resamples. (mc, mc2)}
\item{alpha}{A value of the significant level. (mc, mc2)}
\item{cv}{A value of the coefficient variation. (mc, mc2)}
\item{lhc}{A logical indicates resamples are drawn by Latin hypercude.
(mc, mc2)}
\item{seed}{An integer for the seed of random numbers. (mc, mc2)}
}
\description{
Perform the KS test for the estimated software reliability models
}
\examples{
data(dacs)
result <- fit.srm.nhpp(time=sys1[sys1>=0],
                       te=-sys1[sys1<0], srm.name=c("exp"))
ks.srm.test(result)
results <- fit.srm.nhpp(time=sys1[sys1>=0],
                        te=-sys1[sys1<0], srm.name=c("exp", "gamma"))
sapply(results, function(x) x$p.value)
}
