% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstest.R
\name{ks.srm.test.y}
\alias{ks.srm.test.y}
\title{KS test for software reliability models}
\usage{
ks.srm.test.y(srm, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{srm}{An object for the estimated software reliability model by Rsrat}

\item{alternative}{A string indicates the alternative hypothesis and
must be one of "two.sided" (default), "less", or "greater".}
}
\value{
A list with components;
\item{statistic}{A value of the test statistic.}
\item{p.value.lower}{A lower bound of the p-value of the test.}
\item{p.value.upper}{A upper bound of the p-value of the test.}
\item{alternative}{A string of the alternative hypothesis.}
}
\description{
Peform the KS test for the estimated software reliability models
}
\examples{
data(dacs)
result <- fit.srm.nhpp(fault=sys1g[1:30], srm.name=c("exp"))
ks.srm.test.y(result$srm)
}
