% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KSdistance_group2}
\alias{KSdistance_group2}
\title{KS distance for grouped samples drawn from a continuous distribution.
(experimental: We do not know this is mathematically correct or not.)}
\usage{
KSdistance_group2(ctime, count, cdf)
}
\arguments{
\item{ctime}{A sequence represents time slots (bins)}

\item{count}{A sequence indicates the number of samples falls int a bin}

\item{cdf}{A function of CDF. It is allowd to get a closure.}
}
\value{
A list with components;
\item{dplus}{A value of statistic D+.}
\item{dminus}{A value of statistic D-.}
\item{d}{A value of statistic max(D-,D+).}
}
\description{
KS distance for grouped samples drawn from a continuous distribution.
(experimental: We do not know this is mathematically correct or not.)
}
