% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KSdistance_groupM}
\alias{KSdistance_groupM}
\title{KS distance for grouped samples drawn from a continuous distribution.}
\usage{
KSdistance_groupM(ctime, size, sample, cdf)
}
\arguments{
\item{ctime}{A sequence represents time slots (bins)}

\item{size}{An integer for the number of total samples.}

\item{sample}{A set of sampels drawn from the multinomial distribution.}

\item{cdf}{A function of CDF. It is allowd to get a closure.}
}
\value{
A list with components;
\item{dplus}{A vector of statistic D+.}
\item{dminus}{A vector of statistic D-.}
\item{d}{A vector of statistic max(D-,D+).}
}
\description{
KS distance for grouped samples drawn from a continuous distribution.
}
