% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KSdistance_point}
\alias{KSdistance_point}
\title{KS distance for IID sampes drawn from a continuous distribution.}
\usage{
KSdistance_point(sample, cdf)
}
\arguments{
\item{sample}{A numeric vector of IID samples from a continuous distribution}

\item{cdf}{A function of CDF. It is allowd to get a closure.}
}
\value{
A list with components;
\item{dplus}{A value of statistic D+.}
\item{dminus}{A value of statistic D-.}
\item{d}{A value of statistic max(D-,D+).}
}
\description{
KS distance for IID sampes drawn from a continuous distribution.
}
