% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KSdistance_pointM}
\alias{KSdistance_pointM}
\title{KS distance for IID sampes drawn from a continuous distribution.}
\usage{
KSdistance_pointM(msample, cdf)
}
\arguments{
\item{msample}{A numeric matrix of IID samples from a continuous distribution (m series)}

\item{cdf}{A function of CDF. It is allowd to get a closure.}
}
\value{
A list with components;
\item{dplus}{A value of statistic D+.}
\item{dminus}{A value of statistic D-.}
\item{d}{A value of statistic max(D-,D+).}
}
\description{
KS distance for IID sampes drawn from a continuous distribution.
}
