% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multinomialLHC}
\alias{multinomialLHC}
\title{Generate samples drawn from multinomial distirbution.}
\usage{
multinomialLHC(n, size, ctime, cdf, lhc)
}
\arguments{
\item{n}{An integer indicating the number of samples.}

\item{size}{An integer indicating of dimenisions?}

\item{ctime}{A vector indicating the time instants}

\item{cdf}{An object of cumulative distribution function}

\item{lhc}{A numeric matrix generated by LHC.}
}
\description{
This is a function to generate samples drawn from multinomial distribution fron
a given random sequence of LHC.
}
