% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstest.R
\name{resample.LHC}
\alias{resample.LHC}
\title{Resampling for the grouped data with Latin hypercube}
\usage{
resample.LHC(n, size, ctime, cdf)
}
\arguments{
\item{n}{An integer for the number of samples}

\item{size}{An integer indicates the number of total faults}

\item{ctime}{A sequence for time slots}

\item{cdf}{A function object of the truncated CDF}
}
\value{
A matrix of resamples. Each column corresponds to a sample
}
\description{
Resampling for the grouped data with Latin hypercube
}
