% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{thinning}
\alias{thinning}
\title{Generate samples drawn from NHPP}
\usage{
thinning(intensity, range, maximum)
}
\arguments{
\item{intensity}{A function of intensity function}

\item{range}{A vector (min, max) indicating the range}

\item{maximum}{A value of maximum intensity.}
}
\value{
A time series of event occurences
}
\description{
This is a function to generate samples drwan from NHPP by thining.
}
