% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstest.R
\name{trunc.cdf}
\alias{trunc.cdf}
\title{Truncated CDF for software reliability model}
\usage{
\method{trunc}{cdf}(srm)
}
\arguments{
\item{srm}{An object for the estimated software reliability model}
}
\value{
A function object for the truncated CDF.
}
\description{
Generate the truncated CDF from the estiamted software
reliability model
}
