% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{eic.srm.group}
\alias{eic.srm.group}
\title{EIC for SRM}
\usage{
eic.srm.group(obj, bsample = 100)
}
\arguments{
\item{obj}{An object of an estimated result by Rsrat}

\item{bsample}{An integer of the number of bootstrap samples}
}
\value{
A numeric vector of samples of bias
}
\description{
Generate samples of bias for SRM with group data from bootstrap sampling
}
