% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{eic.srm.sample}
\alias{eic.srm.sample}
\title{EIC for SRM}
\usage{
eic.srm.sample(obj, bsample = 100, initialize = FALSE)
}
\arguments{
\item{obj}{An object of an estimated result by Rsrat}

\item{bsample}{An integer of the number of bootstrap samples}

\item{initialize}{A logical. If it is TRUE, the parameters are reset by initial guesses.
The default is FALSE.}
}
\value{
A vector of samples
}
\description{
Generate samples of EIC
}
